Fix constant memory overflow in CUDA transport
Problem:
- CudaPartitionCache (~387KB) exceeded CUDA constant memory limit (64KB)
- Caused: "ptxas error: File uses too much global constant data"

Solution:
- Move partition function cache from __constant__ to global memory
- Use device pointer (d_partition_ptr) instead of constant symbol
- Allocate with cudaMalloc and set pointer via cudaMemcpyToSymbol
- Update partition_function_lookup() to use pointer dereference

Also fixed:
- atomicAdd type mismatch (int64_t* -> unsigned long long* casts)
- Downbranch data field names (emission_lines, emission_line_count)
- Makefile nvcc linker flags (-Xlinker instead of -Wl,)

Verified:
- test_cuda compiles and runs successfully
- Downbranch table: 7,410,491 entries uploaded to GPU
- MACROATOM mode forced (line_interaction_type = 2)

+++ cuda_atomic.h
@@ -252,7 +252,7 @@ typedef struct {
 
     /* Per-line offsets into emission arrays */
     int64_t  *emission_start;      /* [n_lines] Start index */
-    int32_t  *emission_count;      /* [n_lines] Number of emission options */
+    int64_t  *emission_count;      /* [n_lines] Number of emission options */
 
     /* Emission line data (flattened) */
     int64_t  *emission_line_id;    /* Target line index */
